geodata_pipeline/requirements.txt
rasterio>=1.3
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
laspy[lazrs]>=2.5
open3d>=0.17
pyproj>=3.5
shapely>=2.0
geopandas>=0.13
scikit-image>=0.21
pandas>=2.0
tqdm>=4.65
seaborn>=0.12
cupy-cuda12x>=13.0
